+++ test/message_enricher_test.rb
@@ -2,10 +2,22 @@ require 'test_helper'
 
 class MessageEnricherTest < Test::Unit::TestCase
 
-  def test_enriches_initial_message_when_body_is_passed
+  def test_enriches_initial_message_when_body_contains_error
     response = enriched_response(422, 'InitialMessage', { error: 'My Error' })
 
     assert_equal 'InitialMessage (My Error)', response.message
+  end
+
+  def test_enriches_initial_message_when_body_contains_errors_array
+    response = enriched_response(422, 'InitialMessage', { errors: ['My Error1', 'My Error2'] })
+
+    assert_equal 'InitialMessage (My Error1; My Error2)', response.message
+  end
+
+  def test_enriches_initial_message_when_body_contains_errors_single_value
+    response = enriched_response(422, 'InitialMessage', { errors: 'My Error1' })
+
+    assert_equal 'InitialMessage (My Error1)', response.message
   end
 
   def test_returns_initial_message_when_code_is_200